Harmonize elasticsearch error formating btw. storageEngine and internalEngine (#940)
* Factorize elasticsearch error formating btw. storageEngine and internalEngine
Fix #900

* sonarcube
