+++ lib/api/controllers/funnelController.js
@@ -210,7 +210,9 @@ class FunnelController {
 
     this.checkRights(request)
       .then(modifiedRequest => this.processRequest(modifiedRequest))
-      .then(processResult => callback(null, processResult))
+      .then(processResult => {
+        callback(null, processResult);
+      })
       .catch(err => this._executeError(err, request, true, callback));
 
     return 0;
