+++ lib/api/kuzzle.js
@@ -133,8 +133,8 @@ class Kuzzle extends EventEmitter {
       .then(() => this.services.init())
       .then(() => this.indexCache.init())
       .then(() => this.gc.init())
+      .then(() => this.pluginsManager.trigger('log:info', 'Services initiated'))
       .then(() => {
-        this.pluginsManager.trigger('log:info', 'Services initiated');
         this.funnel.init();
         this.router.init();
         this.statistics.init();
