+++ lib/services/elasticsearch.js
@@ -22,9 +22,9 @@
 'use strict';
 
 const
-  debug = require('../kuzzleDebug')('kuzzle:services:elasticsearch'),
   _ = require('lodash'),
   Bluebird = require('bluebird'),
+  ESCommon = require('../util/esCommon'),
   Service = require('./service'),
   Request = require('kuzzle-common-objects').Request,
   es = require('elasticsearch'),
@@ -33,72 +33,11 @@ const
   {
     BadRequestError,
     InternalError: KuzzleInternalError,
-    ServiceUnavailableError,
     NotFoundError,
-    KuzzleError,
-    ExternalServiceError,
     PreconditionError
   } = require('kuzzle-common-objects').errors;
 
 const scrollCachePrefix = '_docscroll_';
-
-const errorMessagesMapping = [
-  {
-    // [illegal_argument_exception] object mapping [titi] can't be changed from nested to non-nested
-    regex: /^\[illegal_argument_exception] object mapping \[(.*?)] can't be changed from nested to non-nested$/,
-    replacement: 'Can not change mapping for field "$1" from nested to another type'
-  },
-  {
-    // [illegal_argument_exception] object mapping [baz] can't be changed from non-nested to nested
-    regex: /^\[illegal_argument_exception] object mapping \[(.*?)] can't be changed from non-nested to nested$/,
-    replacement: 'Can not change mapping for field "$1" from object to another type'
-  },
-  {
-    // [illegal_argument_exception] Can't merge a non object mapping [aeaze] with an object mapping [aeaze]
-    regex: /^\[illegal_argument_exception] Can't merge a non object mapping \[(.*?)] with an object mapping \[(.*?)]$/,
-    replacement: 'Can not change mapping for field "$1" from object to a scalar type'
-  },
-  {
-    // [illegal_argument_exception] [tutu.tutu] is defined as an object in mapping [aze] but this name is already used for a field in other types
-    regex: /^\[illegal_argument_exception] \[(.*?)] is defined as an object in mapping \[(.*?)] but this name is already used for a field in other types$/,
-    replacement: 'Can not set mapping for field "$1" on collection "$2" because the field name is already used in another collection with a different type'
-  },
-  {
-    // [illegal_argument_exception] mapper [source.flags] of different type, current_type [string], merged_type [long]
-    regex: /^\[illegal_argument_exception] mapper \[(.*?)] of different type, current_type \[(.*?)], merged_type \[(.*?)]$/,
-    replacement: 'Can not change type of field "$1" from "$2" to "$3"'
-  },
-  {
-    // [mapper_parsing_exception] Mapping definition for [flags] has unsupported parameters:  [index : not_analyzed]
-    // eslint-disable-next-line no-regex-spaces
-    regex: /^\[mapper_parsing_exception] Mapping definition for \[(.*?)] has unsupported parameters: \[(.*?)]$/,
-    replacement: 'Parameter "$2" is not supported for field "$1"'
-  },
-  {
-    // [mapper_parsing_exception] No handler for type [booleasn] declared on field [not]
-    regex: /^\[mapper_parsing_exception] No handler for type \[(.*?)] declared on field \[(.*?)]$/,
-    replacement: 'Can not set mapping for field "$2" because type "$1" does not exist'
-  },
-  {
-    // [mapper_parsing_exception] failed to parse [conditions.host.flags]
-    regex: /^\[mapper_parsing_exception] failed to parse \[(.*?)]$/,
-    replacement: 'Failed to validate value of field "$1". Are you trying to insert nested value in a non-nested field ?'
-  },
-  {
-    // [index_not_found_exception] no such index, with { resource.type=index_or_alias & resource.id=foso & index=foso }
-    regex: /^\[index_not_found_exception] no such index, with { resource\.type=([^\s]+) (& )?resource\.id=([^\s]+) (& )?(index_uuid=.* )?index=([^\s]+) }$/,
-    replacement: 'Index "$3" does not exist, please create it first'
-  },
-  {
-    // [mapper_parsing_exception] Expected map for property [fields] on field [foo] but got a class java.lang.String
-    regex: /^\[mapper_parsing_exception] Expected map for property \[fields] on field \[(.*?)] but got a class java\.lang\.String$/,
-    replacement: 'Mapping for field "$1" must be an object with a property "type"'
-  },
-  {
-    regex: /^\[version_conflict_engine_exception] \[data]\[(.*?)]: version conflict.*$/,
-    replacement: 'Unable to modify document "$1": cluster sync failed (too many simultaneous changes applied)'
-  }
-];
 
 /**
  * @property {Kuzzle} kuzzle
@@ -176,7 +115,7 @@ class ElasticSearch extends Service {
         response.nodes = res.nodes;
         return response;
       })
-      .catch(error => Bluebird.reject(this.formatESError(error)));
+      .catch(error => Bluebird.reject(ESCommon.formatESError(error)));
   }
 
   /**
@@ -206,7 +145,7 @@ class ElasticSearch extends Service {
       })
       .then(() => this.client.scroll(esRequest))
       .then(result => flattenSearchResults(result))
-      .catch(error => Bluebird.reject(this.formatESError(error)));
+      .catch(error => Bluebird.reject(ESCommon.formatESError(error)));
   }
 
   /**
@@ -239,7 +178,7 @@ class ElasticSearch extends Service {
 
         return flattened;
       })
-      .catch(error => Bluebird.reject(this.formatESError(error)));
+      .catch(error => Bluebird.reject(ESCommon.formatESError(error)));
   }
 
   /**
@@ -266,7 +205,7 @@ class ElasticSearch extends Service {
 
         return prepareMetadata(result);
       })
-      .catch(error => Bluebird.reject(this.formatESError(error)));
+      .catch(error => Bluebird.reject(ESCommon.formatESError(error)));
   }
 
   /**
@@ -295,7 +234,7 @@ class ElasticSearch extends Service {
 
         return result;
       })
-      .catch(error => Bluebird.reject(this.formatESError(error)));
+      .catch(error => Bluebird.reject(ESCommon.formatESError(error)));
   }
 
   /**
@@ -312,7 +251,7 @@ class ElasticSearch extends Service {
       addActiveFilter(esRequest);
     }
     return this.client.count(esRequest)
-      .catch(error => Bluebird.reject(this.formatESError(error)));
+      .catch(error => Bluebird.reject(ESCommon.formatESError(error)));
   }
 
   /**
@@ -358,7 +297,7 @@ class ElasticSearch extends Service {
 
                 return this.refreshIndexIfNeeded(esRequest, _.extend(res, {_source: request.input.body}));
               })
-              .catch(error => Bluebird.reject(this.formatESError(error)));
+              .catch(error => Bluebird.reject(ESCommon.formatESError(error)));
           }
 
           // The document exits and is active, we reject to prevent the creation
@@ -370,7 +309,7 @@ class ElasticSearch extends Service {
             // The document doesn't exist, we create it
             return this.client.create(esRequest)
               .then(result => this.refreshIndexIfNeeded(esRequest, _.extend(result, {_source: request.input.body})))
-              .catch(error => Bluebird.reject(this.formatESError(error)));
+              .catch(error => Bluebird.reject(ESCommon.formatESError(error)));
           }
 
           // A "real" error occurred, we reject it
@@ -380,7 +319,7 @@ class ElasticSearch extends Service {
 
     return this.client.index(esRequest)
       .then(result => this.refreshIndexIfNeeded(esRequest, _.extend(result, {_source: request.input.body})))
-      .catch(error => Bluebird.reject(this.formatESError(error)));
+      .catch(error => Bluebird.reject(ESCommon.formatESError(error)));
   }
 
   /**
@@ -416,7 +355,7 @@ class ElasticSearch extends Service {
 
     return this.client.index(esRequest)
       .then(result => this.refreshIndexIfNeeded(esRequest, _.extend(result, {_source: request.input.body})))
-      .catch(error => Bluebird.reject(this.formatESError(error)));
+      .catch(error => Bluebird.reject(ESCommon.formatESError(error)));
   }
 
   /**
@@ -455,7 +394,7 @@ class ElasticSearch extends Service {
 
     return this.client.update(esRequest)
       .then(result => this.refreshIndexIfNeeded(esRequest, result))
-      .catch(error => Bluebird.reject(this.formatESError(error)));
+      .catch(error => Bluebird.reject(ESCommon.formatESError(error)));
   }
 
   /**
@@ -537,7 +476,7 @@ class ElasticSearch extends Service {
 
     return this.client.update(esRequest)
       .then(result => this.refreshIndexIfNeeded(esRequest, result))
-      .catch(error => Bluebird.reject(this.formatESError(error)));
+      .catch(error => Bluebird.reject(ESCommon.formatESError(error)));
   }
 
   /**
@@ -574,7 +513,7 @@ class ElasticSearch extends Service {
 
         return this.client.bulk(esRequestBulk)
           .then(() => this.refreshIndexIfNeeded(esRequestBulk, {ids}))
-          .catch(error => Bluebird.reject(this.formatESError(error)));
+          .catch(error => Bluebird.reject(ESCommon.formatESError(error)));
       });
   }
 
@@ -610,7 +549,7 @@ class ElasticSearch extends Service {
           }
           return this.client.bulk(esRequestBulk)
             .then(() => this.refreshIndexIfNeeded(esRequestBulk, {ids}))
-            .catch(error => reject(this.formatESError(error)));
+            .catch(error => reject(ESCommon.formatESError(error)));
         });
       });
   }
@@ -633,7 +572,7 @@ class ElasticSearch extends Service {
     }
 
     return this.client.indices.putMapping(esRequest)
-      .catch(error => Bluebird.reject(this.formatESError(error)));
+      .catch(error => Bluebird.reject(ESCommon.formatESError(error)));
   }
 
   /**
@@ -659,7 +598,7 @@ class ElasticSearch extends Service {
     deleteRequest.input.args.refresh = request.input.args.refresh || false;
 
     return this.deleteByQuery(deleteRequest)
-      .catch(error => Bluebird.reject(this.formatESError(error)));
+      .catch(error => Bluebird.reject(ESCommon.formatESError(error)));
   }
 
   /**
@@ -741,7 +680,7 @@ class ElasticSearch extends Service {
 
         return result;
       })
-      .catch(err => Bluebird.reject(this.formatESError(err)));
+      .catch(err => Bluebird.reject(ESCommon.formatESError(err)));
   }
 
   /**
@@ -756,7 +695,7 @@ class ElasticSearch extends Service {
     esRequest.body = request.input.body;
 
     return this.client.indices.putMapping(esRequest)
-      .catch(error => Bluebird.reject(this.formatESError(error)));
+      .catch(error => Bluebird.reject(ESCommon.formatESError(error)));
   }
 
   /**
@@ -780,7 +719,7 @@ class ElasticSearch extends Service {
 
         return Bluebird.reject(new NotFoundError(`No mapping for index "${request.input.resource.index}"`));
       })
-      .catch(error => Bluebird.reject(this.formatESError(error)));
+      .catch(error => Bluebird.reject(ESCommon.formatESError(error)));
   }
 
   /**
@@ -802,7 +741,7 @@ class ElasticSearch extends Service {
 
         return {collections: {stored: collections}};
       })
-      .catch(error => Bluebird.reject(this.formatESError(error)));
+      .catch(error => Bluebird.reject(ESCommon.formatESError(error)));
   }
 
   /**
@@ -820,7 +759,7 @@ class ElasticSearch extends Service {
 
     return this.client.indices.delete({index: deletedIndexes})
       .then(() => ({deleted: deletedIndexes}))
-      .catch(error => Bluebird.reject(this.formatESError(error)));
+      .catch(error => Bluebird.reject(ESCommon.formatESError(error)));
   }
 
   /**
@@ -833,7 +772,7 @@ class ElasticSearch extends Service {
       .then(result => ({
         indexes: Object.keys(result).filter(indexName => indexName !== '' && indexName[0] !== '%')
       }))
-      .catch(error => Bluebird.reject(this.formatESError(error)));
+      .catch(error => Bluebird.reject(ESCommon.formatESError(error)));
   }
 
   /**
@@ -846,7 +785,7 @@ class ElasticSearch extends Service {
     const index = getIndex(request);
 
     return this.client.indices.create({index})
-      .catch(error => Bluebird.reject(this.formatESError(error)));
+      .catch(error => Bluebird.reject(ESCommon.formatESError(error)));
   }
 
   /**
@@ -861,7 +800,7 @@ class ElasticSearch extends Service {
     delete this.settings.autoRefresh[index];
 
     return this.client.indices.delete({index})
-      .catch(error => Bluebird.reject(this.formatESError(error)));
+      .catch(error => Bluebird.reject(ESCommon.formatESError(error)));
   }
 
   /**
@@ -877,7 +816,7 @@ class ElasticSearch extends Service {
     const index = getIndex(request);
 
     return this.client.indices.refresh({index})
-      .catch(error => Bluebird.reject(this.formatESError(error)));
+      .catch(error => Bluebird.reject(ESCommon.formatESError(error)));
   }
 
   /**
@@ -888,7 +827,7 @@ class ElasticSearch extends Service {
     const esRequest = initESRequest(request);
 
     return this.client.indices.exists(esRequest)
-      .catch(error => Bluebird.reject(this.formatESError(error)));
+      .catch(error => Bluebird.reject(ESCommon.formatESError(error)));
   }
 
   /**
@@ -899,7 +838,7 @@ class ElasticSearch extends Service {
     const esRequest = initESRequest(request);
 
     return this.client.indices.existsType(esRequest)
-      .catch(error => Bluebird.reject(this.formatESError(error)));
+      .catch(error => Bluebird.reject(ESCommon.formatESError(error)));
   }
 
   /**
@@ -932,76 +871,6 @@ class ElasticSearch extends Service {
 
     return this.saveSettings()
       .then(() => this.getAutoRefresh(request));
-  }
-
-  /**
-   * Transforms raw ES errors into a normalized Kuzzle version
-   *
-   * @param {Error} error
-   * @returns {KuzzleError}
-   */
-  formatESError(error) {
-    let
-      kuzzleError,
-      messageReplaced,
-      message = error.message || '';
-
-    if (error instanceof KuzzleError) {
-      return error;
-    }
-
-    if (error instanceof es.errors.NoConnections) {
-      return new ServiceUnavailableError('Elasticsearch service is not connected');
-    }
-
-    messageReplaced = errorMessagesMapping.some(mapping => {
-      message = message.replace(mapping.regex, mapping.replacement);
-      return message !== error.message;
-    });
-
-    switch (error.displayName) {
-      case 'BadRequest':
-        if (!messageReplaced) {
-          if (error.body && error.body.error) {
-            message = error.body.error.root_cause ? error.body.error.root_cause[0].reason : error.body.error.reason;
-          }
-
-          debug('unhandled "BadRequest" elasticsearch error: %a', error);
-        }
-
-        kuzzleError = new BadRequestError(message);
-        break;
-      case 'NotFound':
-        if (!messageReplaced) {
-          if (error.body && error.body.error) {
-            message = error.body.error
-              ? error.body.error.reason + ': ' + error.body.error['resource.id']
-              : error.message + ': ' + error.body._id;
-          }
-
-          debug('unhandled "NotFound" elasticsearch error: %a', error);
-        }
-
-        kuzzleError = new NotFoundError(message);
-        break;
-      case 'Conflict':
-        if (!messageReplaced) {
-          debug('unhandled "Conflict" elasticsearch error: %a', error);
-        }
-
-        kuzzleError = new ExternalServiceError(message);
-        break;
-      default:
-        kuzzleError = new ExternalServiceError(message);
-
-        debug('unhandled default elasticsearch error: %a', message);
-        break;
-    }
-
-    kuzzleError.internalError = error;
-    kuzzleError.service = 'elasticsearch';
-
-    return kuzzleError;
   }
 
   /**
