+++ lib/services/internalEngine/index.js
@@ -44,6 +44,7 @@
 'use strict';
 
 const
+  ESCommon = require('../../util/esCommon'),
   _ = require('lodash'),
   Bluebird = require('bluebird'),
   ms = require('ms'),
@@ -51,9 +52,7 @@ const
   semver = require('semver'),
   {
     NotFoundError,
-    KuzzleError,
     InternalError: KuzzleInternalError,
-    ServiceUnavailableError
   } = require('kuzzle-common-objects').errors;
 
 
@@ -149,7 +148,7 @@ class InternalEngine {
 
         return result;
       })
-      .catch(error => handleError(error));
+      .catch(error => Bluebird.reject(ESCommon.formatESError(error)));
   }
 
   /**
@@ -187,7 +186,7 @@ class InternalEngine {
 
         return result;
       })
-      .catch(err => handleError(err));
+      .catch(error => Bluebird.reject(ESCommon.formatESError(error)));
   }
 
   /**
@@ -201,7 +200,7 @@ class InternalEngine {
       type,
       id
     })
-      .catch(error => handleError(error));
+      .catch(error => Bluebird.reject(ESCommon.formatESError(error)));
   }
 
   /**
@@ -219,7 +218,7 @@ class InternalEngine {
     };
 
     return this.client.get(request)
-      .catch(error => handleError(error));
+      .catch(error => Bluebird.reject(ESCommon.formatESError(error)));
   }
 
   /**
@@ -250,7 +249,7 @@ class InternalEngine {
 
         return result;
       })
-      .catch(error => handleError(error));
+      .catch(error => Bluebird.reject(ESCommon.formatESError(error)));
   }
 
   /**
@@ -278,7 +277,7 @@ class InternalEngine {
     // When we write in ES, the response doesn't contain the initial document content
     return this.client.create(request)
       .then(result => _.extend(result, {_source: content}))
-      .catch(error => handleError(error));
+      .catch(error => Bluebird.reject(ESCommon.formatESError(error)));
   }
 
   /**
@@ -306,7 +305,7 @@ class InternalEngine {
     // When we write in ES, the response from it doesn't contain the initial document content
     return this.client.index(request)
       .then(result => _.extend(result, {_source: content}))
-      .catch(error => handleError(error));
+      .catch(error => Bluebird.reject(ESCommon.formatESError(error)));
   }
 
   /**
@@ -333,7 +332,7 @@ class InternalEngine {
     }
 
     return this.client.update(request)
-      .catch(error => handleError(error));
+      .catch(error => Bluebird.reject(ESCommon.formatESError(error)));
   }
 
   /**
@@ -368,7 +367,7 @@ class InternalEngine {
         return Bluebird.reject(new NotFoundError(`Document with id "${id}" not found.`));
       })
       .then(result => _.extend(result, {_source: content}))
-      .catch(error => handleError(error));
+      .catch(error => Bluebird.reject(ESCommon.formatESError(error)));
   }
 
   /**
@@ -391,7 +390,7 @@ class InternalEngine {
     }
 
     return this.client.delete(request)
-      .catch(error => handleError(error));
+      .catch(error => Bluebird.reject(ESCommon.formatESError(error)));
   }
 
   /**
@@ -402,7 +401,7 @@ class InternalEngine {
   listIndexes () {
     return this.client.indices.getMapping()
       .then(response => Object.keys(response))
-      .catch(error => handleError(error));
+      .catch(error => Bluebird.reject(ESCommon.formatESError(error)));
   }
 
   /**
@@ -419,7 +418,7 @@ class InternalEngine {
 
         return [];
       })
-      .catch(error => handleError(error));
+      .catch(error => Bluebird.reject(ESCommon.formatESError(error)));
   }
 
   /**
@@ -429,7 +428,7 @@ class InternalEngine {
    */
   getMapping (data) {
     return this.client.indices.getMapping(data)
-      .catch(error => handleError(error));
+      .catch(error => Bluebird.reject(ESCommon.formatESError(error)));
   }
 
   /**
@@ -446,7 +445,7 @@ class InternalEngine {
           });
         }
       })
-      .catch(error => handleError(error));
+      .catch(error => Bluebird.reject(ESCommon.formatESError(error)));
   }
 
   /**
@@ -458,7 +457,7 @@ class InternalEngine {
     return this.client.indices.delete({
       index: this.index
     })
-      .catch(error => handleError(error));
+      .catch(error => Bluebird.reject(ESCommon.formatESError(error)));
   }
 
   /**
@@ -476,7 +475,7 @@ class InternalEngine {
     };
 
     return this.client.indices.putMapping(request)
-      .catch(error => handleError(error));
+      .catch(error => Bluebird.reject(ESCommon.formatESError(error)));
   }
 
   /**
@@ -487,24 +486,8 @@ class InternalEngine {
     return this.client.indices.refresh({
       index: this.index
     })
-      .catch(error => handleError(error));
+      .catch(error => Bluebird.reject(ESCommon.formatESError(error)));
   }
-}
-
-function handleError(error) {
-  if (error instanceof es.errors.NoConnections) {
-    return Bluebird.reject(new ServiceUnavailableError('es service is not connected'));
-  }
-
-  if (error instanceof KuzzleError) {
-    return Bluebird.reject(error);
-  }
-
-  if (error.status === 404) {
-    return Bluebird.reject(new NotFoundError(`Not found: ${error.body._type}/${error.body._id}`));
-  }
-
-  return Bluebird.reject(new KuzzleInternalError(error));
 }
 
 module.exports = InternalEngine;
